Fixed lesson editing

+++ app/Http/Controllers/LessonController.php
@@ -42,6 +42,7 @@ class LessonController extends Controller
      */
     public function store(Request $request)
     {
+
         $validator = Validator::make($request->all(), [
             'name' => 'required',
             'content' => 'required',
@@ -118,27 +119,46 @@ class LessonController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'name' => 'required',
-            'content' => 'required',
+            'content' => 'required'
         ]);
         if ($validator->fails())  {
             return response()->json(['errors' => $validator->getMessageBag()->toArray()], 422);
         }
-        $lesson = Lesson::findOrFail($id);
+        $lesson = Lesson::findOrfail($id);
         $lesson->content = $request->input('content');
         $lesson->name = $request->input('name');
-        $course = $lesson->course;
-        if($course->user_id != auth()->id()){
+
+        if($lesson->course->user_id != auth()->id()){
             return response()->json([
                 'message' => 'Unauthorized'
             ]);
         }
+
         $lesson->save();
 
+        if($file = $request->file('file')){
+
+            if($file = $lesson->file) {
+                Storage::delete($file->file_path);
+                $lesson->file()->delete();
+            }
+            $file = $request->file('file');
+            $ext = $file->getClientOriginalExtension();
+            $fileName = md5(time()). ".$ext";
+            $path = $file->storePublicly('lessons');
+            $file = new File();
+            $file->file_name = $fileName;
+            $file->file_path = $path;
+
+            $lesson->file()->save($file);
+        }
+
         return response()->json([
-            'message' => 'Successfully updated education!',
+            'message' => 'Successfully updated lesson!',
             'lesson' => $lesson->makeHidden('course')
         ]);
     }
+
 
     /**
      * Remove the specified resource from storage.
@@ -154,7 +174,9 @@ class LessonController extends Controller
                 'message' => 'Unauthorized'
             ]);
         }
-        Storage::delete($lesson->file->file_path);
+        if($lesson->file) {
+            Storage::delete($lesson->file->file_path);
+        }
         $lesson->file()->delete();
         $lesson->delete();
     }
